Use Text ID for bookmark actions
Before, the book_id and page number were were being used to find the appropriate
Text when editing and deleting bookmarks. Using the Text ID directly simplifies
these queries.

Note:
With the current page slider, it is inconvenient to access and update the current
Text.TxID of the current page. So the "Add bookmark" action from the slideout still
uses the `book_id` and `page_num` to access the appropriate Text record when adding
a new bookmark.