Refresh stale bookstats without page reload
Lute's main page fails to reload bookstats from a cached page.

Currently bookstats are only refreshed when a client requests the main page at `/`.
When a book is opened, it's bookstats are deleted such that when a user navigates to a
cached version of the main page (e.g. they use their browser's back button) the
necessary bookstats record is not refreshed.

To remediate this, the `refresh_stats` call has been relocated to a new route at
`/refresh_stale_stats` which is called by the root `index.html` to refresh stats whether
it has been cached or not.