Issue 531: term reference search should handle casing correctly.
This change adds sentence.SeTextLC.  The field is set to '*' if lowercasing the SeText using
the Language.parser returns the same value from sqlite's built-in LOWER function.  This
special character ends up saving many mb of file size in my db, as 99.95% of the sentences
can be downcased with sqlite when searching for references.

Also adds a data cleanup job for users upgrading from prior version.